Create README.md
Winger Ipsum is a project for creating a Lorem Ipsum text generator using the speeches and quips of Jeff Winger, a fictional character on NBC's "Community."  it is almost ready to go, but I'm having trouble getting my JavaScript working.  The website will take input from the user, run it through the code, and a (hidden by default) text box will appear with generated code that can be copied and pasted by the user.  The js file is well commented, but the jQuery file may or may not work (version 1.11.0 minified).  Image and generated text copyright Sony Television, all code is my own.